flags generic and used in hat and art7
