+++ frontend/src/assets/scss/_custom.scss
@@ -957,3 +957,7 @@ body:not(.aside-menu-lg-show) .form-wrapper {
 body {
 	background: #eee;
 }
+
+.muted {
+	color: #aaa;
+}
