+++ frontend/src/components/art7/dataDefinitions/form.js
@@ -1,7 +1,7 @@
 import { intro_fields } from './questionnaire_fields'
 import tab_sub_info from '@/components/common/dataDefinitions/tab_sub_info'
 import tab_attachments from '@/components/common/dataDefinitions/tab_attachments'
-import flags from '@/components/common/dataDefinitions/flags'
+import tab_flags from '@/components/common/dataDefinitions/tab_flags'
 
 const form = {
 	formDetails: {
@@ -827,38 +827,7 @@ const form = {
 				quantity_emitted: null
 			}
 		},
-		flags: {
-			name: 'flags',
-			hasAssideMenu: false,
-			endpoint_url: 'submission_flags_url',
-			status: null,
-			title: 'Flags',
-			titleHtml: '<b>Flags</b>',
-			tooltipHtml: '',
-			detailsHtml: 'Flags',
-			form_fields: flags,
-			get fields_order() {
-				return Object.keys(this.form_fields)
-			},
-			default_properties: {
-				flag_provisional: false,
-				flag_valid: null,
-				flag_superseded: false,
-				flag_checked_blanks: false,
-				flag_has_blanks: false,
-				flag_confirmed_blanks: false,
-				flag_has_reported_a1: false,
-				flag_has_reported_a2: false,
-				flag_has_reported_b1: false,
-				flag_has_reported_b2: false,
-				flag_has_reported_b3: false,
-				flag_has_reported_c1: false,
-				flag_has_reported_c2: false,
-				flag_has_reported_c3: false,
-				flag_has_reported_e: false,
-				flag_has_reported_f: false
-			}
-		}
+		flags: tab_flags
 	}
 }
 export default form
